module_adapter: modules: Change module entry point type to uintptr_t
Change type of the entry_point parameter in the native_system_agent_start
and system_agent_start functions to uintptr_t which allows safe pointer
passing.

+++ src/include/sof/audio/module_adapter/library/native_system_agent.h
@@ -20,7 +20,7 @@ struct native_system_agent {
 	uint32_t module_size;
 };
 
-int native_system_agent_start(uint32_t entry_point, uint32_t module_id, uint32_t instance_id,
+int native_system_agent_start(uintptr_t entry_point, uint32_t module_id, uint32_t instance_id,
 			      uint32_t core_id, uint32_t log_handle, void *mod_cfg, void **iface);
 
 #endif /* __NATIVE_SYSTEM_AGENT_H__ */
